mysql替换成pg，自增id替换成分布式id

+++ xxl-job-admin/src/main/java/com/xxl/job/admin/core/route/strategy/ExecutorRouteConsistentHash.java
@@ -56,7 +56,7 @@ public class ExecutorRouteConsistentHash extends ExecutorRouter {
         return truncateHashCode;
     }
 
-    public String hashJob(int jobId, List<String> addressList) {
+    public String hashJob(long jobId, List<String> addressList) {
 
         // ------A1------A2-------A3------
         // -----------J1------------------
